# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: LMS API
  version: 1.0.0
servers:
  - url: https://api.seudominio.com
paths:
...
  /api/admin/modules/{id}:
    put:
      summary: Atualiza título/ordem do módulo
      security: [{ bearerAuth: [] }]
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                order: { type: integer }
      responses:
        '200': { description: OK }
        '400': { description: Erro de validação }
        '404': { description: Módulo não encontrado }
  /api/admin/modules/{id}/reorder:
    patch:
      summary: Reordena os itens do módulo
      security: [{ bearerAuth: [] }]
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemIds]
              properties:
                itemIds:
                  type: array
                  items: { type: string, format: uuid }
      responses:
        '200': { description: OK }
        '400': { description: Erro de validação ou item fora do módulo }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...